pyproject.toml
[tool.poetry]
name = "sports"
version = "0.1.0"
description = ""
authors = ["leehjhjhj <tsukiakarii@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pymysql = "^1.1.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"
dependency-injector = "^4.41.0"
boto3 = "^1.28.81"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
pytest = "^8.0.1"
pytest-django = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
